+++ controllers/users.controller.js
@@ -1,5 +1,5 @@
-import asyncDec from 'utils/asyncDecorator';
-import userCore from 'core/users';
+import asyncDec from 'utils/asyncDecoration';
+import userCore from 'core/users.core';
 
 const getUserById = asyncDec(async (req, res) => {
   const dataResponse = await userCore.findUserById(req.params.userId);
