+++ routes/auth.js
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import authController from 'controllers/auth';
+import authController from 'controllers/auth.controller';
 
 const router = Router();
